Merge Handlers into UI-Updates

+++ src/main/java/edu/brown/cs/drawbridge/carpools/TripSearcher.java
@@ -140,8 +140,9 @@ public class TripSearcher {
    * @return A List of valid paths. Each path is a List of Trips.
    */
   public List<List<Trip>> searchWithId(String userId, double startLat,
-      double startLon, double endLat, double endLon, int departureTime,
-      double distanceRadius, int timeRadius) {
+      double startLon, double endLat, double endLon, long departureTime,
+      double distanceRadius, double timeRadius) {
+    
     List<List<Trip>> paths = search(userId, startLat, startLon, endLat, endLon,
         departureTime, distanceRadius, timeRadius);
     Collections.sort(paths, TRIP_COMPARATOR);
@@ -171,8 +172,9 @@ public class TripSearcher {
    * @return A List of valid paths. Each path is a List of Trips.
    */
   public List<List<Trip>> searchWithoutId(double startLat, double startLon,
-      double endLat, double endLon, int departureTime, double distanceRadius,
-      int timeRadius) {
+      double endLat, double endLon, long departureTime, double distanceRadius,
+      double timeRadius) {
+
     List<List<Trip>> paths = search("", startLat, startLon, endLat, endLon,
         departureTime, distanceRadius, timeRadius);
     Collections.sort(paths, new CostComparator());
@@ -204,8 +206,9 @@ public class TripSearcher {
    * @return A List of valid paths. Each path is a List of Trips.
    */
   private List<List<Trip>> search(String userId, double startLat,
-      double startLon, double endLat, double endLon, int departureTime,
-      double distanceRadius, int timeRadius) {
+      double startLon, double endLat, double endLon, long departureTime,
+      double distanceRadius, double timeRadius) {
+
     // Set the user id for all comparators
     setUser(userId);
